refactor: split PendingTxs component
